requirements.txt
Flask
gunicorn
python-dotenv
PyPDF2
langchain
langchain-core
langchain-community
langchain-google-genai
langchain-text-splitters
pydantic
pydantic-settings
faiss-cpu